+++ src/clojure/nrepl/middleware/session.clj
@@ -4,6 +4,7 @@
   (:require
    clojure.main
    clojure.test
+   nrepl.middleware.truncate-outputs
    [nrepl.middleware :refer [set-descriptor!]]
    [nrepl.middleware.interruptible-eval :refer [*msg*]]
    [nrepl.misc :refer [uuid response-for]]
@@ -197,7 +198,7 @@
             (h msg)))))))
 
 (set-descriptor! #'session
-                 {:requires #{}
+                 {:requires #{#'nrepl.middleware.truncate-outputs/truncate-outputs}
                   :expects #{}
                   :describe-fn (fn [{:keys [session] :as describe-msg}]
                                  (when (and session (instance? clojure.lang.Atom session))