+++ test/clojure/nrepl/core_test.clj
@@ -229,6 +229,138 @@
               (map :out)
               (remove nil?)))))
 
+(def-repl-test truncate-eval-values-and-outputs
+  (is (= [{:value "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-value "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19"
+           :truncated "true"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-value 50}
+                                :code "(range 30)"})
+              (map #(dissoc % :id :session :ns))))
+      "value size bigger than truncate")
+  (is (= [{:value ":a"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-value 50}
+                                :code ":a"})
+              (map #(dissoc % :id :session :ns))))
+      "value size less than truncate")
+  (is (= [{:out "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-out "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19"
+           :truncated "true"}
+          {:value "nil"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-output 50}
+                                :code "(print (range 30))"})
+              (map #(dissoc % :id :session :ns))))
+      "out should be truncated")
+  (is (= [{:ex "class java.lang.NullPointerException",
+           :root-ex "class java.lang.NullPointerException",
+           :status ["eval-error"]}
+          {:err "NullPointerException   clojure.lang.Numbers.ops (Numbers.java:1018)\n"
+           :truncated-err "NullPointerE"
+           :truncated "true"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-output 12}
+                                :code "(+ 1 nil)"})
+              (map #(dissoc % :id :session :ns))))
+      "err should be truncated")
+  (is (= [{:ex "class java.lang.NullPointerException",
+           :root-ex "class java.lang.NullPointerException",
+           :status ["eval-error"]}
+          {:err "NullPointerException   clojure.lang.Numbers.ops (Numbers.java:1018)\n"
+           :truncated-err "NullPointe"
+           :truncated "true"}
+          {:value "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-value "(0 1 2"
+           :truncated "true"}
+          {:out "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-out "(0 1 2 3 4"
+           :truncated "true"}
+          {:value "nil"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-value 6
+                                                   :truncate-output 10}
+                                :code "(+ 1 nil) (range 30) (print (range 30))"})
+              (map #(dissoc % :id :session :ns))))
+      "value, out and err should all be truncated")
+  (is (= [{:ex "class java.lang.NullPointerException",
+           :root-ex "class java.lang.NullPointerException",
+           :status ["eval-error"]}
+          {:err "NullPointerException   clojure.lang.Numbers.ops (Numbers.java:1018)\n"}
+          {:value "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49)"}
+          {:out "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49)"}
+          {:value "nil"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {}
+                                :code "(+ 1 nil) (range 50) (print (range 50))"})
+              (map #(dissoc % :id :session :ns))))
+      "nothing should be truncated when not passing :truncate-value or :truncate-output to the message"))
+
+(defn custom-truncate-strategy
+  [v {:keys [truncate-init truncate]}]
+  (if (or (nil? truncate) (<= (count v) truncate))
+    nil
+    (subs v truncate-init truncate)))
+
+(def-repl-test truncation-with-custom-truncate-fns
+  (is (= [{:value "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-value "2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19"
+           :truncated "true"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-init 5
+                                                   :truncate 50}
+                                :truncate-value-strategy `custom-truncate-strategy
+                                :code "(range 30)"})
+              (map #(dissoc % :id :session :ns))))
+      "custom truncate strategy for value")
+  (is (= [{:out "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-out "2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19"
+           :truncated "true"}
+          {:value "nil"}
+          {:ex "class java.lang.NullPointerException",
+           :root-ex "class java.lang.NullPointerException",
+           :status ["eval-error"]}
+          {:err "NullPointerException   clojure.lang.Numbers.ops (Numbers.java:1018)\n",
+           :truncated "true",
+           :truncated-err "ointerException   clojure.lang.Numbers.ops (N"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-init 5
+                                                   :truncate 50}
+                                :truncate-output-strategy `custom-truncate-strategy
+                                :code "(print (range 30)) (+ 5 nil)"})
+              (map #(dissoc % :id :session :ns))))
+      "custom truncate strategy for outputs")
+  (is (= [{:out "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-out "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19"
+           :truncated "true"}
+          {:value "nil"}
+          {:ex "class java.lang.NullPointerException",
+           :root-ex "class java.lang.NullPointerException",
+           :status ["eval-error"]}
+          {:err "NullPointerException   clojure.lang.Numbers.ops (Numbers.java:1018)\n",
+           :truncated "true",
+           :truncated-err "NullPointerException   clojure.lang.Numbers.ops (N"}
+          {:value "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29)"
+           :truncated-value "(0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19"
+           :truncated "true"}
+          {:status ["done"]}]
+         (->> (message session {:op :eval
+                                :truncate-options {:truncate-output 50
+                                                   :truncate-value 50}
+                                :truncate-value-strategy `my.missing.ns/truncate-value
+                                :truncate-output-strategy `my.missing.ns/truncate-output
+                                :code "(print (range 30)) (+ 5 nil) (range 30)"})
+              (map #(dissoc % :id :session :ns))))
+      "bad symbols should fall back to default truncate strategies"))
+
 (def-repl-test ensure-whitespace-prints
   (is (= " \t \n \f \n" (->> (repl-eval client "(println \" \t \n \f \")")
                              combine-responses