+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   function in `pr-values`, enabling pretty-printed REPL results.
 * [#66](https://github.com/nrepl/nrepl/pull/66): Add support for a global and local configuration file.
 * [#63](https://github.com/nrepl/nrepl/issues/63): Make it possible to specify the transport via the command-line client (`--transport/-t`).
+* [#75](https://github.com/nrepl/nrepl/issues/75): Make it possible to limit the size of returned values and output.
 
 #### Bugs fixed